# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API-Jira server.",
        "version": "0.0.1",
        "title": "Jira-API",
        "termsOfService": ""
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/get/jira/getOBjectinType": {
            "get": {
                "tags": [
                    "Get Jira"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "12"
                    },
                    {
                        "name": "ObjectTypeId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "HKNOC-Customer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/objectschema/list": {
            "get": {
                "tags": [
                    "Get Jira"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/workflow/list": {
            "get": {
                "tags": [
                    "Get Jira"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sql/getTableData": {
            "get": {
                "tags": [
                    "Get SQL"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "dsbs"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "abc123"
                    },
                    {
                        "name": "connString",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "172.25.15.138:1521/fbiu"
                    },
                    {
                        "name": "tableName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "DSBS.JIRA_CPE_TYPE"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/get/jira/object/objectsWithNamesAttributes": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Asset Management"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "CPE Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/set/jira/issue/createIssue": {
            "post": {
                "tags": [
                    "Create Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "createIssue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issueAttribute"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/set/jira/issue/createIssueByName": {
            "post": {
                "tags": [
                    "Create Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "createIssueByName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issueAttributeByName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/set/jira/issue/updateIssue": {
            "post": {
                "tags": [
                    "Create Jira Issue"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "updateIssue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issueAttributeUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/get/jira/ExportCSVreport": {
            "get": {
                "tags": [
                    "Get Jira"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "projectKey",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "IBVSD"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "issueAttribute": {
            "type": "object",
            "properties": {
                "createIssue": {
                    "type": "object",
                    "properties": {
                        "project": {
                            "type": "object",
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "enum": "AM"
                                }
                            }
                        },
                        "issuetype": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "enum": "Computer Asset"
                                }
                            }
                        },
                        "summary": {
                            "type": "string",
                            "enum": "Created by API"
                        },
                        "customfield_12317": {
                            "type": "string",
                            "enum": "77790211"
                        },
                        "customfield_12318": {
                            "type": "string",
                            "enum": "910040611"
                        },
                        "customfield_12320": {
                            "type": "string",
                            "enum": "D12-345678"
                        },
                        "customfield_12321": {
                            "type": "string",
                            "enum": "1809-008841"
                        },
                        "customfield_12322": {
                            "type": "string",
                            "enum": "FGL1839263S"
                        },
                        "customfield_12327": {
                            "type": "string",
                            "enum": "MACROVIEW,RENT"
                        },
                        "customfield_12328": {
                            "type": "string",
                            "enum": "SI quotation no.: QN/2018/07245/TH"
                        },
                        "customfield_12329": {
                            "type": "string",
                            "enum": "Macroview"
                        },
                        "customfield_12330": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "customfield_12331": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "customfield_12332": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "ROUTER"
                                }
                            }
                        },
                        "customfield_12333": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "CISCO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "issueAttributeByName": {
            "type": "object",
            "properties": {
                "createIssueByName": {
                    "type": "object",
                    "properties": {
                        "project": {
                            "type": "object",
                            "properties": {
                                "key": {
                                    "type": "string",
                                    "enum": "AM"
                                }
                            }
                        },
                        "issuetype": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "enum": "Computer Asset"
                                }
                            }
                        },
                        "Summary": {
                            "type": "string",
                            "enum": "Created by API"
                        },
                        "Contract No.": {
                            "type": "string",
                            "enum": "77790211"
                        },
                        "Circuit No.": {
                            "type": "string",
                            "enum": "910040611"
                        },
                        "Order No.": {
                            "type": "string",
                            "enum": "D12-345678"
                        },
                        "BCase No.": {
                            "type": "string",
                            "enum": "1809-008841"
                        },
                        "Serial No.": {
                            "type": "string",
                            "enum": "FGL1839263S"
                        },
                        "SI_EQUIPMENT": {
                            "type": "string",
                            "enum": "MACROVIEW,RENT"
                        },
                        "SI_INFO": {
                            "type": "string",
                            "enum": "SI quotation no.: QN/2018/07245/TH"
                        },
                        "SI_NAME": {
                            "type": "string",
                            "enum": "Macroview"
                        },
                        "Maintenance Start Date": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "Maintenance End Date": {
                            "type": "string",
                            "enum": "2011-10-03"
                        },
                        "CPE Type": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "ROUTER"
                                }
                            }
                        },
                        "Manufacturer": {
                            "type": "object",
                            "properties": {
                                "value": {
                                    "type": "string",
                                    "enum": "CISCO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "issueAttributeUpdate": {
            "type": "object",
            "properties": {
                "updateIssue": {
                    "example": {
                        "issueId": "IBVSD-9",
                        "fields": {
                            "customfield_12020": "TEST"
                        }
                    }
                }
            }
        },
...
    }
}